atualização

+++ api/src/main/java/br/com/alunoonline/api/service/DisciplinaService.java
@@ -2,6 +2,10 @@ package br.com.alunoonline.api.service;
 
 import br.com.alunoonline.api.domain.disciplina.Disciplina;
 import br.com.alunoonline.api.domain.disciplina.DisciplinaRepository;
+import br.com.alunoonline.api.domain.disciplina.MatriculaAlunoRepository;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -14,11 +18,15 @@ import java.util.Optional;
 public class DisciplinaService {
 
     @Autowired
-    DisciplinaRepository disciplinaRepository;
+    private DisciplinaRepository disciplinaRepository;
 
-    public void criarDisciplina (Disciplina disciplina){
+    @Autowired
+    private MatriculaAlunoRepository matriculaAlunoRepository;
+
+    public void criarDisciplina(Disciplina disciplina){
         disciplinaRepository.save(disciplina);
     }
+
     public List<Disciplina> listarTodasDisciplinas(){
         return disciplinaRepository.findAll();
     }
@@ -28,6 +36,17 @@ public class DisciplinaService {
     }
 
     public void deletarDisciplinaPorId(Long id){
+        // Verifica se existem alunos matriculados
+        boolean possuiAlunosMatriculados = !matriculaAlunoRepository.findByDisciplinaId(id).isEmpty();
+
+        if (possuiAlunosMatriculados) {
+            throw new ResponseStatusException(
+                    HttpStatus.BAD_REQUEST,
+                    "Não é possível deletar a disciplina pois existem alunos matriculados."
+            );
+        }
+
+        // Se não houver alunos, deleta a disciplina
         disciplinaRepository.deleteById(id);
     }
 
@@ -37,6 +56,7 @@ public class DisciplinaService {
         if (disciplinaDoBancoDeDados.isEmpty()) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Disciplina não encontrada no banco de dados.");
         }
+
         Disciplina disciplinaParaEditar = disciplinaDoBancoDeDados.get();
         disciplinaParaEditar.setNome(disciplina.getNome());
         disciplinaParaEditar.setCargaHoraria(disciplina.getCargaHoraria());
@@ -46,6 +66,5 @@ public class DisciplinaService {
 
     public List<Disciplina> listarDisciplinasDoProf(Long professorId){
         return disciplinaRepository.findByProfessorId(professorId);
-
     }
 }
